Back instalado y funcionando.

+++ Front/Angular Personas V3 (Guards)/angular-personas/src/app/components/person-delete/person-delete.component.ts
@@ -9,6 +9,7 @@ import { ActivatedRoute } from '@angular/router';
 import {
   Location
 } from '@angular/common';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-person-delete',
@@ -35,8 +36,25 @@ export class PersonDeleteComponent implements OnInit {
   }
   delete(): void {
     const id = Number(this.route.snapshot.paramMap.get('id'));
-    this.personService.deletePerson(id).subscribe();
-    this.goBack();
+    Swal.fire({
+      title: '¿Estás seguro de querer eliminar al usuario Nº: '+id+'?',
+      text: "Si confirmas, tendrás que crearlo de nuevo.",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Si, entiendo los riesgos'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.personService.deletePerson(id).subscribe();
+        Swal.fire(
+          'Usuario Nº: '+id+' borrado satisfactoriamente.',
+          '',
+          'success'
+        ).then(function(){window.location.href="localhost:4200/personas"});
+      }
+      this.goBack();
+    })
   }
   goBack(): void {
     this.location.back();
